backend/requirements.txt
# TraumaInsight Backend Dependencies
fastapi==0.134.0
uvicorn[standard]==0.41.0
sqlalchemy[asyncio]==2.0.47
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.8.1
python-dotenv==1.2.1
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==0.26.0
aiosqlite==0.21.0
eval-type-backport==0.3.1
